indexer/requirements.txt
-i https://pypi.org/simple
aiosqlite==0.20.0; python_version >= '3.8'
alembic==1.13.2; python_version >= '3.8'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.4.0; python_version >= '3.8'
apprise==1.9.0; python_version >= '3.6'
asgi-lifespan==2.1.0; python_version >= '3.7'
async-timeout==4.0.3; python_full_version < '3.12.0'
asyncpg==0.29.0; python_full_version >= '3.8.0'
attrs==24.2.0; python_version >= '3.7'
cachetools==5.5.0; python_version >= '3.7'
certifi==2024.8.30; python_version >= '3.6'
cffi==1.17.0; platform_python_implementation != 'PyPy'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
cloudpickle==3.0.0; python_version >= '3.8'
colorama==0.4.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
coolname==2.2.0
croniter==2.0.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
cryptography==43.0.0; python_version >= '3.7'
dateparser==1.2.0; python_version >= '3.7'
dnspython==2.6.1; python_version >= '3.8'
docker==7.1.0; python_version >= '3.8'
email-validator==2.2.0
exceptiongroup==1.2.2; python_version >= '3.7'
fsspec==2024.6.1; python_version >= '3.8'
google-auth==2.34.0; python_version >= '3.7'
graphviz==0.20.3; python_version >= '3.8'
greenlet==3.0.3
griffe==1.2.0; python_version >= '3.8'
h11==0.14.0; python_version >= '3.7'
h2==4.1.0
hpack==4.0.0; python_full_version >= '3.6.1'
httpcore==1.0.5; python_version >= '3.8'
httpx[http2]==0.27.2; python_version >= '3.8'
humanize==4.10.0; python_version >= '3.8'
hyperframe==6.0.1; python_full_version >= '3.6.1'
idna==3.8; python_version >= '3.6'
importlib-resources==6.4.4; python_version >= '3.8'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.4; python_version >= '3.7'
jinja2-humanize-extension==0.4.0; python_version >= '3.0'
jsonpatch==1.33; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
jsonpointer==3.0.0; python_version >= '3.7'
jsonschema==4.23.0; python_version >= '3.8'
jsonschema-specifications==2023.12.1; python_version >= '3.8'
kubernetes==29.0.0; python_version >= '3.6'
mako==1.3.5; python_version >= '3.8'
markdown==3.7; python_version >= '3.8'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==2.1.5; python_version >= '3.7'
mdurl==0.1.2; python_version >= '3.7'
oauthlib==3.2.2; python_version >= '3.6'
orjson==3.10.7; python_version >= '3.8'
packaging==24.1; python_version >= '3.8'
pathspec==0.12.1; python_version >= '3.8'
pendulum==2.1.2; python_version < '3.12'
prefect==2.20.4; python_version >= '3.8'
prefect-docker==0.5.4; python_version >= '3.8'
pyasn1==0.6.0; python_version >= '3.8'
pyasn1-modules==0.4.0; python_version >= '3.8'
pycparser==2.22; python_version >= '3.8'
pydantic[email]==2.8.2; python_version >= '3.8'
pydantic-core==2.20.1; python_version >= '3.8'
pygments==2.18.0; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1; python_version >= '3.8'
python-multipart==0.0.9; python_version >= '3.8'
python-slugify==8.0.4; python_version >= '3.7'
pytz==2024.1
pytzdata==2020.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyyaml==6.0.2; python_version >= '3.8'
readchar==4.2.0; python_version >= '3.8'
referencing==0.35.1; python_version >= '3.8'
regex==2024.7.24; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
requests-oauthlib==2.0.0; python_version >= '3.4'
rfc3339-validator==0.1.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
rich==13.8.0; python_full_version >= '3.7.0'
rpds-py==0.20.0; python_version >= '3.8'
rsa==4.9; python_version >= '3.6' and python_version < '4'
ruamel.yaml==0.18.6; python_version >= '3.7'
ruamel.yaml.clib==0.2.8; python_version < '3.13' and platform_python_implementation == 'CPython'
shellingham==1.5.4; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy[asyncio]==2.0.32; python_version >= '3.7'
text-unidecode==1.3
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
typer==0.12.5; python_version >= '3.7'
typing-extensions==4.12.2; python_version >= '3.8'
tzlocal==5.2; python_version >= '3.8'
ujson==5.10.0; python_version >= '3.8'
urllib3==2.2.2; python_version >= '3.8'
uvicorn==0.30.6; python_version >= '3.8'
websocket-client==1.8.0; python_version >= '3.8'
websockets==13.0.1; python_version >= '3.8'
